+++ src/extension.ts
@@ -8,6 +8,7 @@
 
 import * as vscode from 'vscode';
 import * as engine from './engine';
+import crypto from 'node:crypto';
 
 /* ******************************************************
  *
@@ -21,17 +22,22 @@ import * as engine from './engine';
  *
  */
 
+// ********************
+// CODICE DI PROVA. VA CANCELLATO PRIMA DEL RILASCIO
+//
 const text =
   'Lorem Ipsum è un testo segnaposto utilizzato nel settore della tipografia e della stampa.';
 const pw = 'sammarino';
 
 const encryptCurrentFile = async () => {
-  const tk: string | undefined = await engine.encrypted(text, pw);
-  const ta: string | undefined = await engine.decrypted(tk, pw);
+  const tk: string | undefined = await engine.encrypt(text, pw);
+  const ta: string | undefined = await engine.decrypt(tk, pw);
   console.log('§>', { text, tk, ta });
 };
 
 const decryptCurrentFile = () => {};
+//
+// *********************
 
 export function activate(context: vscode.ExtensionContext) {
   let disposableEncrypt = vscode.commands.registerCommand(
@@ -53,3 +59,5 @@ export function activate(context: vscode.ExtensionContext) {
 }
 
 export function deactivate() {}
+
+// End of file extension.ts